+++ src/tier.hpp
@@ -28,8 +28,8 @@ namespace fs = boost::filesystem;
 
 class Tier{
 public:
-	unsigned long watermark_bytes;
-  unsigned long pinned_files_size;
+	unsigned long long watermark_bytes;
+  unsigned long long pinned_files_size;
 	int watermark;
 	fs::path dir;
 	std::string id;
@@ -37,8 +37,8 @@ public:
 	Tier(std::string id_){
 		id = id_;
 	}
-	unsigned long get_capacity();
-	unsigned long get_usage();
+	unsigned long long get_capacity();
+	unsigned long long get_usage();
 	void cleanup(void){
 		incoming_files.erase(incoming_files.begin(), incoming_files.end());
 	}
